+++ src/anyio/abc/_tasks.py
@@ -6,7 +6,15 @@ from types import TracebackType
 from typing import TYPE_CHECKING, Any, Generic, TypeVar
 
 if TYPE_CHECKING:
+    from mypy_extensions import NamedArg, VarArg
+    from trio_typing import takes_callable_and_args
+
     from .._core._tasks import CancelScope
+else:
+
+    def takes_callable_and_args(fn):
+        return fn
+
 
 T_Retval = TypeVar("T_Retval")
 
@@ -32,10 +40,12 @@ class TaskGroup(metaclass=ABCMeta):
     cancel_scope: CancelScope
 
     @abstractmethod
+    @takes_callable_and_args
     def start_soon(
         self,
-        func: Callable[..., Coroutine[Any, Any, Any]],
-        *args: object,
+        func: Callable[..., Coroutine[Any, Any, Any]]
+        | Callable[[VarArg()], Coroutine[Any, Any, Any]],
+        *args: Any,
         name: object = None,
     ) -> None:
         """
@@ -49,12 +59,17 @@ class TaskGroup(metaclass=ABCMeta):
         """
 
     @abstractmethod
+    @takes_callable_and_args
     async def start(
         self,
-        func: Callable[..., Awaitable[Any]],
-        *args: object,
+        func: Callable[..., Awaitable[Any]]
+        | Callable[
+            [VarArg(), NamedArg(TaskStatus[T_Retval], "task_status")],  # noqa: F821
+            Awaitable[Any],
+        ],
+        *args: Any,
         name: object = None,
-    ) -> object:
+    ) -> T_Retval:
         """
         Start a new task and wait until it signals for readiness.